[FIX] models: generalize the protection mechanism to not recompute some fields
The protection prevents fields to be invalidated/recomputed on some records.
This mechanism is useful against accidental cache invalidation when computing
or inversing fields.  It also prevents to trigger the recomputation of fields
before and/or after their inversion.

Add a test on stored field with compute and inverse methods: the test checks
how many times the compute and inverse methods are invoked in several cases.
Essentially, the field should not be recomputed when it is written.