[FIX] account: remove related payments upon statement line cancellation.
Since commit 3d48f2a375f22fc3da0f013041cfe0ce9ba6fb13, each statement line creates a payment entry if it creates a move. So the code to cancel a statement line must be adapted to keep the old behavior: remove the payment if it was created by the statement line processing; keep the payment if it was registered before the statement line.