[FIX] web_tour: tips activation when a modal is open
When a modal is open, we only activate tips that are inside the modal.

The former solution didn't work when the tip's selector contains ','
(often used when the selector is different in community and enterprise,
e.g. selectors for apps), because the modal restriction was only
applied on the first part of the selector.