[FIX] website: correct design of welcome message
The welcome message was using the mt/mb classes to add some space on
top and bottom of it. This is fine when the background-color of the
website is white, but this is not the case with every theme.

To fix the problem, add padding in LESS instead. Also move the inline
style in LESS to prevent theme extension problems.