[FIX] point_of_sale: missing tax in accounting entry
The accounting item generated from a PoS order does not contain the
related taxes. Although the accounting entry is globally correct since
the tax line is created manually, the tax should still be written on the
accounting item for reporting purpose.

Note that this is already the behavior when the accounting entry is
generated directly from an invoice.

Fixes #10855