[FIX] ir_qweb: issue if multiple IRA are created for a bundle
If multiple transactions start knowing that they should generate the
bundle, each one will run lessc and create a ir.attachment record
referencing the path on the filestore. While the file is not duplicated
on the filestore (as it is stored according to its hash), the final
result of `t-call-assets` will source multiple time the same file.

This patch fixes this issue by adapting get_attachment to group its
results by file name and only select the one with the max id (latest
generated).

We also remove the `inc` argument as it is not used.