[FIX] point_of_sale: display the correct UOM in the ScaleScreenWidget
The scale screen widget was hardcoded to show kilograms as the
UOM. This makes it so that the UOM set on the product being weighed is
used instead.

Note that this bug was purely visual. It did not convert the weight to
the wrong UOM or anything like that.

See also d043fd03ceca9935d910fb6e8393025a11984b1b.

opw-684374