[FIX] web: set correct web client height
The main content of the web client was set to grow/shrink so that it
fills the window (thanks to flex). However, to be able to shrink as
much as necessary the min-height needs to be forced to 0.

Without this fix, the web client was overflowing on the bottom of the
page, making impossible to access some menu items.