[FIX] website: /website/info use website.info xmlid
When a user try to create a new page info, he is redirected
to the /page/info because the xmlid already exists...

The problem is that this view need to be rendered via the
controller /website/info which gives some extra values (version, ...)

The fix for v8 is to redirect /page/info to /website/info.

In V9, the behavior is changed and one page info-1 will be created.
But this fix is also required for direct access to /page/info.

To do in master: maybe use another xml_id that this basic name 'info'?

This commit fix issue #8022