[IMP] calendar: notification route should be in auth
The 2 calendar notification routes were in auth
`none`. This causes no problem, but since they are
reimplementing the auth user mecanism, it is better
to set them directly in auth `user`. Moreover, they
are only called from webclient, when the user is
logged.

This commit also migrate them to new API (we forgot
them ....)