[IMP] point_of_sale: create a partner with empty fields
When creating a partner, the fields where the information is not specified were
filled with an empty string value.
The empty string value is not ideal as prevents to use filters like "email not
set" and forces to use more complexe domain "... or email = ''"

Instead, set the value false for the fields were nothing is provided (same
behaviour as backend web client).

Remove the fields.barcode overwrite to empty string as it is probably an
unintentional side effect of 3b33e5c3.

opw-684902