[FIX] website_form: traceback when sending an attachment
The form builder offers the possibility to add a "Custom File Upload"
field. When the user clicks on "Send", an error occurs ("An error has
occured, the form has not been sent.").

This is because we try to send a mail linked to "mail.mail", which
doesn't make sense.

The case was actually taken into account in the code, there was just an
oversight in the code.

opw-684040